requirements.txt
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
natsort==8.4.0
numpy==2.2.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pathspec==0.12.1
protobuf==5.29.3
psutil==7.0.0
Pygments==2.19.1
Pypubsub==4.0.3
PySocks==1.7.1
requests==2.32.3
rich==14.0.0
rich-click==1.8.9
robotframework==7.2.2
robotframework-assertion-engine==3.0.3
robotframework-browser==19.5.1
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.7.1
robotframework-tidy==4.17.0
seedir==0.5.0
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
wxPython==4.2.3
